+++ lib/oli_web/live/users/author_detail_view.ex
@@ -20,11 +20,30 @@ defmodule OliWeb.Users.AuthorsDetailView do
   }
 
   prop author, :any
-  data breadcrumbs, :any, default: [Breadcrumb.new(%{full_title: "Author Details"})]
+  data breadcrumbs, :any
   data title, :string, default: "Author Details"
   data user, :struct, default: nil
   data modal, :any, default: nil
   data csrf_token, :any
+
+  defp set_breadcrumbs(author) do
+    OliWeb.Admin.AdminView.breadcrumb()
+    |> OliWeb.Users.AuthorsView.breadcrumb()
+    |> breadcrumb(author)
+  end
+
+  def breadcrumb(previous, %Author{id: id} = author) do
+    name =
+      OliWeb.Users.UsersTableModel.normalize(author.name, author.given_name, author.family_name)
+
+    previous ++
+      [
+        Breadcrumb.new(%{
+          full_title: name,
+          link: Routes.live_path(OliWeb.Endpoint, __MODULE__, id)
+        })
+      ]
+  end
 
   def mount(
         %{"user_id" => details_id},
@@ -35,11 +54,12 @@ defmodule OliWeb.Users.AuthorsDetailView do
 
     case Repo.get(Author, details_id) do
       nil ->
-        {:ok, redirect(socket, to: Routes.static_page_path(OliWeb.Endpoint, :unauthorized))}
+        {:ok, redirect(socket, to: Routes.static_page_path(OliWeb.Endpoint, :not_found))}
 
       user ->
         {:ok,
          assign(socket,
+           breadcrumbs: set_breadcrumbs(user),
            author: author,
            user: user,
            csrf_token: csrf_token
@@ -133,12 +153,15 @@ defmodule OliWeb.Users.AuthorsDetailView do
         socket
       ) do
     author = Accounts.get_author!(id)
-    {:ok, _author} = Accounts.delete_author(author)
 
-    {:noreply,
-     socket
-     |> assign(user: author)
-     |> hide_modal()}
+    case Accounts.delete_author(author) do
+      {:ok, _} ->
+        {:noreply,
+         redirect(socket, to: Routes.live_path(OliWeb.Endpoint, OliWeb.Users.AuthorsView))}
+
+      {:error, e} ->
+        {:noreply, put_flash(socket, :error, e)}
+    end
   end
 
   def handle_event("show_lock_account_modal", _, socket) do