+++ lib/oli_web/live/users/authors_view.ex
@@ -18,7 +18,7 @@ defmodule OliWeb.Users.AuthorsView do
   }
 
   prop author, :any
-  data breadcrumbs, :any, default: [Breadcrumb.new(%{full_title: "All Authors"})]
+  data breadcrumbs, :any
   data title, :string, default: "All Authors"
 
   data authors, :list, default: []
@@ -28,6 +28,21 @@ defmodule OliWeb.Users.AuthorsView do
   data offset, :integer, default: 0
   data limit, :integer, default: @limit
   data options, :any
+
+  defp set_breadcrumbs() do
+    OliWeb.Admin.AdminView.breadcrumb()
+    |> breadcrumb()
+  end
+
+  def breadcrumb(previous) do
+    previous ++
+      [
+        Breadcrumb.new(%{
+          full_title: "All Authors",
+          link: Routes.live_path(OliWeb.Endpoint, __MODULE__)
+        })
+      ]
+  end
 
   def mount(_, %{"current_author_id" => author_id}, socket) do
     author = Repo.get(Author, author_id)
@@ -45,6 +60,7 @@ defmodule OliWeb.Users.AuthorsView do
 
     {:ok,
      assign(socket,
+       breadcrumbs: set_breadcrumbs(),
        author: author,
        authors: authors,
        total_count: total_count,