+++ lib/oli_web/live/users/user_detail_view.ex
@@ -2,7 +2,7 @@ defmodule OliWeb.Users.UsersDetailView do
   use Surface.LiveView, layout: {OliWeb.LayoutView, "live.html"}
   alias Oli.Repo
   alias OliWeb.Common.Breadcrumb
-  alias Oli.Accounts.Author
+  alias Oli.Accounts.{Author, User}
   alias OliWeb.Common.Properties.{Groups, Group, ReadOnly}
   import OliWeb.Common.Properties.Utils
   alias Oli.Accounts
@@ -19,11 +19,17 @@ defmodule OliWeb.Users.UsersDetailView do
   }
 
   prop author, :any
-  data breadcrumbs, :any, default: [Breadcrumb.new(%{full_title: "User Details"})]
+  data breadcrumbs, :any
   data title, :string, default: "User Details"
   data user, :struct, default: nil
   data modal, :any, default: nil
   data csrf_token, :any
+
+  defp set_breadcrumbs(user) do
+    OliWeb.Admin.AdminView.breadcrumb()
+    |> OliWeb.Users.UsersView.breadcrumb()
+    |> breadcrumb(user)
+  end
 
   def mount(
         %{"user_id" => user_id},
@@ -34,11 +40,12 @@ defmodule OliWeb.Users.UsersDetailView do
 
     case Accounts.get_user_by(id: user_id) do
       nil ->
-        {:ok, redirect(socket, to: Routes.static_page_path(OliWeb.Endpoint, :unauthorized))}
+        {:ok, redirect(socket, to: Routes.static_page_path(OliWeb.Endpoint, :not_found))}
 
       user ->
         {:ok,
          assign(socket,
+           breadcrumbs: set_breadcrumbs(user),
            author: author,
            user: user,
            csrf_token: csrf_token
@@ -179,11 +186,26 @@ defmodule OliWeb.Users.UsersDetailView do
         socket
       ) do
     user = Accounts.get_user!(id)
-    {:ok, _user} = Accounts.delete_user(user)
 
-    {:noreply,
-     socket
-     |> assign(user: user)
-     |> hide_modal()}
+    case Accounts.delete_user(user) do
+      {:ok, _} ->
+        {:noreply,
+         redirect(socket, to: Routes.live_path(OliWeb.Endpoint, OliWeb.Users.UsersView))}
+
+      {:error, e} ->
+        {:noreply, put_flash(socket, :error, e)}
+    end
+  end
+
+  def breadcrumb(previous, %User{id: id} = user) do
+    name = OliWeb.Users.UsersTableModel.normalize(user.name, user.given_name, user.family_name)
+
+    previous ++
+      [
+        Breadcrumb.new(%{
+          full_title: name,
+          link: Routes.live_path(OliWeb.Endpoint, __MODULE__, id)
+        })
+      ]
   end
 end