+++ lib/oli_web/router.ex
@@ -195,6 +195,7 @@ defmodule OliWeb.Router do
 
     get("/", StaticPageController, :index)
     get("/unauthorized", StaticPageController, :unauthorized)
+    get("/not_found", StaticPageController, :not_found)
   end
 
   scope "/", OliWeb do
@@ -632,6 +633,7 @@ defmodule OliWeb.Router do
       :pow_email_layout
     ])
 
+    live("/", Admin.AdminView)
     live("/authors", Users.AuthorsView)
     live("/authors/:user_id", Users.AuthorsDetailView)
     live("/users", Users.UsersView)