+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### Enhancements
 
+- New Admin landing page
+
 ## 0.14.0 (2021-10-13)
 
 ### Bug Fixes
@@ -33,7 +35,6 @@
 - Add ability to remix materials from multiple projects
 - Fix insert content popup in page editor
 - Add blackboard LTI 1.3 configuration instructions
-
 
 ### Release Notes